Change route and billing line item rate type to decimal
Updated the rate parameter in the routes API endpoints and the Route and BillingLineItem models to use BigDecimal instead of Integer. This change allows for fractional rates, ensuring more accurate financial calculations. Adjusted related specs to reflect the new rate format and added validations for decimal rates.

+++ app/api/api/v1/endpoints/routes/create.rb
@@ -17,7 +17,7 @@ module API::V1::Endpoints::Routes
       params do
         requires :origin,      type: String,  documentation: {desc: "The origin of the route."}
         requires :destination, type: String,  documentation: {desc: "The destination of the route."}
-        requires :rate,        type: Integer, documentation: {desc: "The rate of the route."}
+        requires :rate,        type: BigDecimal, documentation: {desc: "The rate of the route."}
       end
       post "/create" do
         authorize!(::Route, :create)
